Tip: install R and RStudio locally anyway.
In `01-introduction.Rmd`, "Tip: This Lesson works better in the cloud" gave information for running the genomics workshop command line VMs, not for the RStudio Server that `genomics-r-intro` actually uses. So, I've proposed changes to this that link instead to the R ecology lesson setup page, and encourage learners to install R and RStudio locally. In workshops that we have had in Edinburgh, there have been instances of a (local) server going down so we cannot continue using studio servers. If my proposed solution is not appropriate, the tip should anyway be updated to be specific to an RStudio Server lesson.